+++ content/docs/deployment/csminstallationwizard/src/static/js/tests/generate-yaml.test.js
@@ -115,6 +115,10 @@ describe("GIVEN setValues function", () => {
             <input type="text" id="manage-array-id">
             <input type="text" id="manage-array-endpoint-url">
             <input type="checkbox" id="topology">
+			<input type="checkbox" id="rename-sdc">
+			<input type="text" id="sdc-prefix" value="">
+			<input type="checkbox" id="approve-sdc">
+			<input type="checkbox" id="enable-quota">
         `;
 
 		const expected = {
@@ -174,7 +178,7 @@ describe("GIVEN setValues function", () => {
 			vSphereVCenterCredSecret: undefined,
 			targetClusterId: "",
 			targetArrayID: "",
-			targetUnisphere: "",
+			targetUnisphere: '""',
 			iscsichap: false,
 			manageArrayId: '""',
 			manageArrayEndpointUrl: '""',
@@ -184,6 +188,10 @@ describe("GIVEN setValues function", () => {
 			observabilityOperatorTopology: false,
 			topologyEnabled: false,
 			transportProtocol: "",
+			renameSDC: false,
+			sdcPrefix: "",
+			approveSDC: false,
+			enableQuota: false,
 		};
 
 		const received = setValues(testCSMMap, CONSTANTS);
@@ -232,6 +240,10 @@ describe("GIVEN setValues function", () => {
             <input type="text" id="manage-array-endpoint-url">
             <input type="checkbox" id="topology">
             <input type="checkbox" id="leader-election">
+			<input type="checkbox" id="rename-sdc">
+			<input type="text" id="sdc-prefix" value="">
+			<input type="checkbox" id="approve-sdc">
+			<input type="checkbox" id="enable-quota">
         `;
 
 		const expected = {
@@ -286,7 +298,7 @@ describe("GIVEN setValues function", () => {
 			portGroups: undefined,
 			targetClusterId: "",
 			targetArrayID: "",
-			targetUnisphere: "",
+			targetUnisphere: '""',
 			vSphereEnabled: false,
 			vSphereFCPortGroup: undefined,
 			vSphereFCHostName: undefined,
@@ -301,6 +313,10 @@ describe("GIVEN setValues function", () => {
 			observabilityOperatorTopology: false,
 			topologyEnabled: false,
 			transportProtocol: "",
+			renameSDC: false,
+			sdcPrefix: "",
+			approveSDC: false,
+			enableQuota: false,
 		};
 
 		const received = setValues(testCSMMap, CONSTANTS);
@@ -441,14 +457,14 @@ describe("GIVEN createYamlString function", () => {
       enabled: $POWERMAX_ENABLED
       global:
         storageArrays:
-          - storageArrayId: ""
-            endpoint: ""
+         - storageArrayId: ""
+           endpoint: ""
             backupEndpoint: ""
-          - storageArrayId: ""
-            endpoint: ""
+         - storageArrayId: ""
+           endpoint: ""
         managementServers:
-          - endpoint: ""
-          - endpoint: ""
+         - endpoint: ""
+         - endpoint: ""
       version: v2.6.0
       images:
         driverRepository: $IMAGE_REPOSITORY
@@ -514,6 +530,11 @@ describe("GIVEN createYamlString function", () => {
         healthMonitor:
           enabled: $HEALTH_MONITOR_ENABLED
         nodeSelector: $NODE_POD_NODE_SELECTOR
+        renameSDC:
+		  enabled: $RENAME_SDC_ENABLED
+		  sdcPrefix: $SDC_PREFIX
+		approveSDC:
+		  enabled: $APPROVE_SDC_ENABLED
         tolerations:
         # Uncomment if CSM for Resiliency and CSI Driver pods monitor is enabled 
         # - key: "offline.vxflexos.storage.dell.com"
@@ -677,7 +698,10 @@ describe("GIVEN createYamlString function", () => {
 		vSphereFCHostName: "csi-vsphere-VC-HN",
 		vSphereVCenterHost: "00.000.000.00",
 		vSphereVCenterCredSecret: "vcenter-creds",
-		migration: false
+		migration: false,
+		renameSDC: false,
+		sdcPrefix: "sdc-test",
+		approveSDC: false
 	};
 
 	test("SHOULD return generated yaml file string for driver csi-powerstore", () => {
@@ -749,14 +773,14 @@ describe("GIVEN createYamlString function", () => {
       enabled: false
       global:
         storageArrays:
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
             backupEndpoint: ""
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
         managementServers:
-          - endpoint: ""
-          - endpoint: ""
+   #      - endpoint: ""
+   #      - endpoint: ""
       version: v2.6.0
       images:
         driverRepository: dellemc
@@ -822,6 +846,11 @@ describe("GIVEN createYamlString function", () => {
         healthMonitor:
           enabled: false
         nodeSelector: false
+        renameSDC:
+		  enabled: false
+		  sdcPrefix: sdc-test
+		approveSDC:
+		  enabled: false
         tolerations:
         # Uncomment if CSM for Resiliency and CSI Driver pods monitor is enabled 
         # - key: "offline.vxflexos.storage.dell.com"
@@ -1018,14 +1047,14 @@ describe("GIVEN createYamlString function", () => {
       enabled: false
       global:
         storageArrays:
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
             backupEndpoint: ""
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
         managementServers:
-          - endpoint: ""
-          - endpoint: ""
+   #      - endpoint: ""
+   #      - endpoint: ""
       version: v2.6.0
       images:
         driverRepository: dellemc
@@ -1091,6 +1120,11 @@ describe("GIVEN createYamlString function", () => {
         healthMonitor:
           enabled: false
         nodeSelector: false
+        renameSDC:
+		  enabled: false
+		  sdcPrefix: sdc-test
+		approveSDC:
+		  enabled: false
         tolerations:
         # Uncomment if CSM for Resiliency and CSI Driver pods monitor is enabled 
         # - key: "offline.vxflexos.storage.dell.com"
@@ -1288,14 +1322,14 @@ describe("GIVEN createYamlString function", () => {
       enabled: true
       global:
         storageArrays:
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
             backupEndpoint: ""
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
         managementServers:
-          - endpoint: ""
-          - endpoint: ""
+   #      - endpoint: ""
+   #      - endpoint: ""
       version: v2.6.0
       images:
         driverRepository: dellemc
@@ -1361,6 +1395,11 @@ describe("GIVEN createYamlString function", () => {
         healthMonitor:
           enabled: false
         nodeSelector: false
+        renameSDC:
+		  enabled: false
+		  sdcPrefix: sdc-test
+		approveSDC:
+		  enabled: false
         tolerations:
         # Uncomment if CSM for Resiliency and CSI Driver pods monitor is enabled 
         # - key: "offline.vxflexos.storage.dell.com"
@@ -1560,14 +1599,14 @@ describe("GIVEN createYamlString function", () => {
       enabled: false
       global:
         storageArrays:
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
             backupEndpoint: ""
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
         managementServers:
-          - endpoint: ""
-          - endpoint: ""
+   #      - endpoint: ""
+   #      - endpoint: ""
       version: v2.6.0
       images:
         driverRepository: dellemc
@@ -1633,6 +1672,11 @@ describe("GIVEN createYamlString function", () => {
         healthMonitor:
           enabled: false
         nodeSelector: false
+        renameSDC:
+		  enabled: false
+		  sdcPrefix: sdc-test
+		approveSDC:
+		  enabled: false
         tolerations:
         # Uncomment if CSM for Resiliency and CSI Driver pods monitor is enabled 
         # - key: "offline.vxflexos.storage.dell.com"
@@ -1831,14 +1875,14 @@ describe("GIVEN createYamlString function", () => {
       enabled: false
       global:
         storageArrays:
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
             backupEndpoint: ""
-          - storageArrayId: ""
-            endpoint: ""
+   #      - storageArrayId: ""
+   #       endpoint: ""
         managementServers:
-          - endpoint: ""
-          - endpoint: ""
+   #      - endpoint: ""
+   #      - endpoint: ""
       version: v2.6.0
       images:
         driverRepository: dellemc
@@ -1904,6 +1948,11 @@ describe("GIVEN createYamlString function", () => {
         healthMonitor:
           enabled: false
         nodeSelector: false
+        renameSDC:
+		  enabled: false
+		  sdcPrefix: sdc-test
+		approveSDC:
+		  enabled: false
         tolerations:
         # Uncomment if CSM for Resiliency and CSI Driver pods monitor is enabled 
         # - key: "offline.vxflexos.storage.dell.com"