+++ content/v1/deployment/csminstallationwizard/release_notes.md
@@ -3,11 +3,11 @@ title: Release Notes
 description: Release notes for CSM Installation Wizard
 ---
 
-## Release Notes - CSM Installation Wizard 1.0.0
+## Release Notes - CSM Installation Wizard 1.0.1
 
 ### New Features/Changes
 
-- [Added support for the CSI-PowerStore, CSI-PowerMax, CSI-PowerScale, CSI-PowerFlex, CSI-Unity and the supported modules for helm installation](https://github.com/dell/csm/issues/698)
+- [Added support for the CSI-PowerStore, CSI-PowerMax, CSI-PowerScale, CSI-PowerFlex, CSI-Unity and the supported modules for helm installation ](https://github.com/dell/csm/issues/698)
 
 ### Fixed Issues